--- src/scss/App.scss
@@ -1,1 +1,0 @@
-@import "grid-kit/scss/grid-kit";
